+++ Changelog.md
@@ -8,6 +8,8 @@
 - automatically zoom to newly imported locations
 - added exporting CZML layers to .kmz files
 - added clustering of Location pins and Layer elements when zooming out
+- keep distance to terrain when zooming to different location pins or the
+  current location; the viewing distance is also re-used after app restarts
 
 ## Version 1.9.0